# remove socket read goroutine; creating garbage, triggering GC and high CPU