I noticed that once statsd or graphite exporters were enabled, CPU utilization would go crazy after a short while. I played around with pprof and noticed high GC calls. The root cause was the go routine to proactively read and discard incoming socket data performing *large* numbers of non-blocking socket reads. This should be unnecessary to keep the writer unblocked.