@jaqx0r 2015-11-13
I just ran the benchmarks after this pull:

before
BenchmarkRsyncdProgram-4	      50	  24429285 ns/op	  12.80 MB/s
BenchmarkSftpProgram-4  	     300	   6155250 ns/op	3491.71 MB/s
BenchmarkDhcpdProgram-4 	       1	8782405473 ns/op	   0.02 MB/s

after
BenchmarkRsyncdProgram-4	     100	  11614059 ns/op	 206.25 MB/s
BenchmarkSftpProgram-4  	     500	   3871499 ns/op	15361.13 MB/s
BenchmarkDhcpdProgram-4 	       1	7545562716 ns/op	   0.03 MB/s

that seems significant.