@jaqx0r 2015-11-13
(op and MB are misleading in these benchmarks, they refer to a full pass of the test data and MB really means lines processed.)