@ryancox 2015-11-13
Hmmm I wouldn't have anticipated that. The benchmarks do a one-shot with no pushTargets. The affected code path is only hit when pushTargets is > 0.

```
// StartMetricPush pushes metrics to the configured services each interval.
func (e *Exporter) StartMetricPush() {
	if len(e.pushTargets) > 0 {
```