@jaqx0r 2015-11-14
Hmm, there is some other oddity happening, then.  Most recent run now:

BenchmarkRsyncdProgram-4	      50	  23636631 ns/op	  13.23 MB/s
BenchmarkSftpProgram-4  	     200	   7499873 ns/op	1594.46 MB/s
BenchmarkDhcpdProgram-4 	       1	9186606961 ns/op	   0.01 MB/s