@jaqx0r 2015-11-14
I was too optimistic there. But these wildy variant benchmarks need some
investigation. Thanks again for the PR.

On Sat, 14 Nov 2015, 09:40 Ryan Cox <notifications@github.com> wrote:

> Hmmm I wouldn't have anticipated that. The benchmarks do a one-shot with
> no pushTargets. The affected code path is only hit when pushTargets is > 0.
>
> // StartMetricPush pushes metrics to the configured services each interval.
> func (e *Exporter) StartMetricPush() {
>     if len(e.pushTargets) > 0 {
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/google/mtail/pull/14#issuecomment-156579643>.
>